+++ app/src/main/java/com/dayrayaneh/automation/view/pishkhanItemView/foroshSakhtAfzar/ForooshSakhtAfzarActivity.java
@@ -344,8 +344,8 @@ public class ForooshSakhtAfzarActivity extends BaseActivity {
     }
 
     private void setTotalInfo(List<DataItem> dataItem) {
-        int mablaqKol = 0;
-        int tedadAqlam = 0;
+        long mablaqKol = 0;
+        long tedadAqlam = 0;
 
         for (int i = 0; i < dataItem.size(); i++) {
             mablaqKol += dataItem.get(i).getTotalPriceSellHard();
